consolidate CDC & PUC
asdf
